refactor: 既存の Repository を legacy に移した
